Add more features
pypi-publish.yml
 - add PYPI CI

tkgui
 - add META_LABEL metadata
 - add metadata_list config editor
 - add rois to plot_widget.listbox
 - refactored nexus_treeview.py and added hdf_viewer.py and nexus_string_views.py
 - added default path to Nexus file viewer

 Experiment
  - add_data_paths function
  - improvement of multi-line plot

 NexusScan
  - moved NexusScan classes into seperate file

 mkdocs.yml
  - removed search to see if this brings search back
