+++ mmg_toolbox/tkguis/apps/nexus.py
@@ -1,3 +1,7 @@
+"""
+
+"""
+
 import tkinter as tk
 
 from mmg_toolbox.tkguis.misc.functions import topmenu, select_hdf_file
@@ -8,10 +12,11 @@ def create_nexus_viewer(filename: str, parent: tk.Misc | None = None) -> RootWit
     """
     File Viewer of the NeXus structure
     """
-    from ..widgets.nexus_treeview import HDFViewer
+    from mmg_toolbox.tkguis.widgets.hdf_viewer import HDFViewer
 
     root = create_root(parent=parent, window_title='NeXus File Viewer')
 
+    #TODO: add config
     widget = HDFViewer(root, filename)
 
     menu = {
