+++ mmg_toolbox/tkguis/widgets/nexus_treeview.py
@@ -10,15 +10,10 @@ from tkinter import ttk
 import hdfmap
 from hdfmap.eval_functions import generate_identifier
 
-from mmg_toolbox.utils.file_functions import hdfobj_string
-from ..misc.styles import update_text_style
-from ..misc.functions import post_right_click_menu, open_close_all_tree, select_hdf_file
-from ..misc.search import search_tree
+from ..misc.functions import post_right_click_menu
 from ..misc.logging import create_logger
 
 logger = create_logger(__file__)
-
-DETAILS_TAB_WIDTH = 30
 
 
 def right_click_menu(frame, tree):
@@ -87,40 +82,6 @@ class _Treeview:
 
     def delete(self):
         self.tree.delete(*self.tree.get_children())
-
-
-class _StringView:
-    """
-    String viewer widget for NeXus file viewer
-    """
-
-    def __init__(self, root: tk.Misc):
-
-        frm = ttk.Frame(root)
-        frm.pack(side=tk.LEFT, expand=tk.YES, fill=tk.BOTH)
-
-        self.text = tk.Text(frm, wrap=tk.NONE)
-
-        vbar = ttk.Scrollbar(frm, orient=tk.VERTICAL, command=self.text.yview)
-        hbar = ttk.Scrollbar(frm, orient=tk.HORIZONTAL, command=self.text.xview)
-        self.text.configure(yscrollcommand=vbar.set, xscrollcommand=hbar.set)
-
-        hbar.pack(side=tk.BOTTOM, fill=tk.X)
-        vbar.pack(side=tk.RIGHT, fill=tk.Y)
-        self.text.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES)
-
-        if hasattr(root, 'style'):
-            update_text_style(self.text, root.style)
-
-    def _populate(self, string):
-        self.delete()
-        self.text.insert('1.0', string)
-
-    def populate(self, **kwargs):
-        pass
-
-    def delete(self):
-        self.text.delete('1.0', tk.END)
 
 
 class HdfTreeview(_Treeview):
@@ -202,7 +163,9 @@ class HdfNameSpace(_Treeview):
         self.tree.heading("path", text="Path")
         self.tree.heading("value", text="Value")
 
-    def populate(self, hdf_obj: h5py.File, hdf_map: hdfmap.NexusMap):
+    def populate(self, hdf_obj: h5py.File, hdf_map: hdfmap.NexusMap,
+                 all: bool = True, group: bool = False, combined: bool = False, values: bool = False,
+                 arrays: bool = False, scannables: bool = False, metadata: bool = False, image_data: bool = False):
         """Load HDF file, populate ttk.treeview object"""
 
         data = {
@@ -210,359 +173,50 @@ class HdfNameSpace(_Treeview):
             for name, path in hdf_map.combined.items()
         }
 
-        datasets = self.tree.insert("", tk.END, text='Groups', values=('', ''))
-        for name, path_list in hdf_map.classes.items():
-            # path_list = list(set(path_list))  # remove duplicates
-            if len(path_list) == 1:
-                self.tree.insert(datasets, tk.END, text=name, values=(path_list[0], ''))
-            else:
-                grp = self.tree.insert(datasets, tk.END, text=name, values=('', ''))
-                for path in path_list:
-                    self.tree.insert(grp, tk.END, text=name, values=(path, ''))
+        if all or group:
+            datasets = self.tree.insert("", tk.END, text='Groups', values=('', ''))
+            for name, path_list in hdf_map.classes.items():
+                # path_list = list(set(path_list))  # remove duplicates
+                if len(path_list) == 1:
+                    self.tree.insert(datasets, tk.END, text=name, values=(path_list[0], ''))
+                else:
+                    grp = self.tree.insert(datasets, tk.END, text=name, values=('', ''))
+                    for path in path_list:
+                        self.tree.insert(grp, tk.END, text=name, values=(path, ''))
+
+        if all or combined:
+            datasets = self.tree.insert("", tk.END, text='Combined', values=('', ''))
+            for name, path in hdf_map.combined.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
+
+        if all or values:
+            datasets = self.tree.insert("", tk.END, text='Values', values=('', ''))
+            for name, path in hdf_map.values.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
+
+        if all or arrays:
+            datasets = self.tree.insert("", tk.END, text='Arrays', values=('', ''))
+            for name, path in hdf_map.arrays.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
+
+        if all or scannables:
+            datasets = self.tree.insert("", tk.END, text='Scannables', values=('', ''))
+            for name, path in hdf_map.scannables.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
+
+        if all or metadata:
+            datasets = self.tree.insert("", tk.END, text='Metadata', values=('', ''))
+            for name, path in hdf_map.metadata.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
+
+        if all or image_data:
+            datasets = self.tree.insert("", tk.END, text='Image Data', values=('', ''))
+            for name, path in hdf_map.image_data.items():
+                value = data.get(name, 'NOT IN MAP')
+                self.tree.insert(datasets, tk.END, text=name, values=(path, value))
 
-        datasets = self.tree.insert("", tk.END, text='Combined', values=('', ''))
-        for name, path in hdf_map.combined.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-        datasets = self.tree.insert("", tk.END, text='Values', values=('', ''))
-        for name, path in hdf_map.values.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-        datasets = self.tree.insert("", tk.END, text='Arrays', values=('', ''))
-        for name, path in hdf_map.arrays.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-        datasets = self.tree.insert("", tk.END, text='Scannables', values=('', ''))
-        for name, path in hdf_map.scannables.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-        datasets = self.tree.insert("", tk.END, text='Metadata', values=('', ''))
-        for name, path in hdf_map.metadata.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-        datasets = self.tree.insert("", tk.END, text='Image Data', values=('', ''))
-        for name, path in hdf_map.image_data.items():
-            value = data.get(name, 'NOT IN MAP')
-            self.tree.insert(datasets, tk.END, text=name, values=(path, value))
-
-
-class HdfTreeStr(_StringView):
-    """
-    HDF Tree String object
-    """
-    def populate(self, hdf_filename: str):
-        """Load HDF file, populate ttk.treeview object"""
-        tree_str = hdfmap.hdf_tree_string(hdf_filename)
-        self._populate(tree_str)
-
-
-class HdfNexusStr(_StringView):
-    """
-    Nexus Info String object
-    """
-
-    def populate(self, hdf_map: hdfmap.NexusMap):
-        """Load HDF file, populate ttk.treeview object"""
-        tree_str = hdf_map.info_nexus()
-        self._populate(tree_str)
-
-
-class Nexus2SrsStr(_StringView):
-    """
-    Display output of NeXus2SRS
-    """
-
-    def populate(self, hdf_map: hdfmap.NexusMap):
-        """Load HDF file, populate ttk.treeview object"""
-        from nexus2srs.nexus2srs import generate_datafile
-        with hdf_map.load_hdf() as hdf:
-            outstr, detector_image_paths = generate_datafile(hdf, hdf_map)
-        self._populate(outstr)
-
-
-class NxTransformationsStr(_StringView):
-    """
-    Display output of NeXus2Transformations
-    """
-    def populate(self, hdf_filename: str):
-        from mmg_toolbox.nexus.nexus_transformations import generate_nxtranformations_string
-        outstr = generate_nxtranformations_string(hdf_filename)
-        self._populate(outstr)
-
-
-############################################################################
-############################ File Viewer ###################################
-############################################################################
-
-
-class HDFViewer:
-    """
-    HDF Viewer - display cascading hierarchical data within HDF file in ttk GUI
-        HDFViewer("filename.h5")
-    Simple ttk interface for browsing HDF file structures.
-     - Click Browse or File>Select File to pick an HDF, H5 or NeXus file
-     - Collapse and expand the tree to view the file structure
-     - Search for addresses using the search bar
-     - Click on a dataset or group to view stored attributes and data
-
-    :param root: tk root
-    :param hdf_filename: str or None*, if str opens this file initially
-    """
-
-    def __init__(self, root: tk.Misc, hdf_filename: str = None):
-        self.map: hdfmap.NexusMap | None = None
-        self.root = root
-
-        # Variables
-        self.filepath = tk.StringVar(self.root, '')
-        self.expandall = tk.BooleanVar(self.root, False)
-        self.expression_box = tk.StringVar(self.root, '')
-        self.expression_path = tk.StringVar(self.root, 'path = ')
-        self.search_box = tk.StringVar(self.root, '')
-        self.search_matchcase = tk.BooleanVar(self.root, False)
-        self.search_wholeword = tk.BooleanVar(self.root, True)
-
-        "------- Build Elements -----"
-        # filepath
-        self.ini_browse(self.root)
-
-        main = ttk.Frame(self.root)
-        main.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-
-        frm = ttk.Frame(main)
-        frm.pack(side=tk.LEFT, expand=tk.YES, fill=tk.BOTH)
-
-        # Tabs
-        self.view_tabs = ttk.Notebook(frm)
-        tab1 = ttk.Frame(self.view_tabs)
-        tab2 = ttk.Frame(self.view_tabs)
-        tab3 = ttk.Frame(self.view_tabs)
-        tab4 = ttk.Frame(self.view_tabs)
-        tab5 = ttk.Frame(self.view_tabs)
-        tab6 = ttk.Frame(self.view_tabs)
-        self.view_tabs.bind('<<NotebookTabChanged>>', self.tab_change)
-
-        self.view_tabs.add(tab1, text='HDF Tree')
-        self.view_tabs.add(tab2, text='NeXus')
-        self.view_tabs.add(tab3, text='HdfMap')
-        self.view_tabs.add(tab4, text='Tree String')
-        self.view_tabs.add(tab5, text='NeXus2SRS')
-        self.view_tabs.add(tab6, text='NXtransformations')
-        self.view_tabs.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES)
-        # treeviews
-        self.hdf_tree = HdfTreeview(tab1)
-        self.nexus = HdfNexusStr(tab2)
-        self.hdf_map = HdfNameSpace(tab3)
-        self.hdf_text = HdfTreeStr(tab4)
-        self.nexus2srs = Nexus2SrsStr(tab5)
-        self.transformations = NxTransformationsStr(tab6)
-
-        self.hdf_tree.tree.bind('<<TreeviewSelect>>', self.tree_select)
-        self.hdf_map.tree.bind('<<TreeviewSelect>>', self.tree_select)
-
-        frm = ttk.Frame(main)
-        frm.pack(side=tk.LEFT, expand=tk.NO, fill=tk.BOTH)
-        # notebook
-        tab_detail, tab_search, tab_expr = self.ini_notebook(frm)
-
-        self.text = self.ini_details(tab_detail)
-        self.ini_search(tab_search)
-        self.text2 = self.ini_expression(tab_expr)
-
-        if hasattr(self.root, 'style'):
-            update_text_style(self.text, self.root.style)
-            update_text_style(self.text2, self.root.style)
-
-        "-------- Start Mainloop ------"
-        if hdf_filename:
-            self.filepath.set(hdf_filename)
-            self.populate_from_file()
-
-    "======================================================"
-    "================= init functions ====================="
-    "======================================================"
-
-    def ini_browse(self, frame: tk.Misc):
-        frm = ttk.Frame(frame)
-        frm.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-
-        var = ttk.Button(frm, text='Browse', command=self.select_file, width=10)
-        var.pack(side=tk.LEFT)
-
-        var = ttk.Entry(frm, textvariable=self.filepath)
-        var.pack(side=tk.LEFT, expand=tk.YES, fill=tk.BOTH)
-        var.bind('<Return>', self.populate_from_file)
-        var.bind('<KP_Enter>', self.populate_from_file)
-
-        var = ttk.Checkbutton(frm, variable=self.expandall, text='Expand', command=self.check_expand)
-        var.pack(side=tk.LEFT)
-
-    def ini_notebook(self, frame: tk.Misc) -> tuple[ttk.Frame, ttk.Frame, ttk.Frame]:
-
-        frm = ttk.Frame(frame)
-        frm.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES)
-
-        tab_control = ttk.Notebook(frm)
-        tab1 = ttk.Frame(tab_control)
-        tab2 = ttk.Frame(tab_control)
-        tab3 = ttk.Frame(tab_control)
-
-        tab_control.add(tab1, text='Details')
-        tab_control.add(tab2, text='Search')
-        tab_control.add(tab3, text='Expression')
-        tab_control.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES)
-
-        return tab1, tab2, tab3
-
-    def ini_details(self, frame: tk.Misc) -> tk.Text:
-        frm = ttk.Frame(frame)
-        frm.pack(side=tk.LEFT, expand=tk.YES, fill=tk.BOTH)
-
-        text = tk.Text(frm, wrap=tk.NONE, width=DETAILS_TAB_WIDTH)
-        text.pack(fill=tk.BOTH, expand=tk.YES)
-
-        var = tk.Scrollbar(frm, orient=tk.HORIZONTAL, command=text.xview)
-        var.pack(side=tk.BOTTOM, fill=tk.X)
-        text.configure(xscrollcommand=var.set)
-        return text
-
-    def ini_search(self, frame: tk.Misc):
-        frm = ttk.Frame(frame)
-        frm.pack(side=tk.TOP, expand=tk.YES, fill=tk.X)
-
-        var = ttk.Entry(frm, textvariable=self.search_box)
-        var.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-        # var.bind('<KeyRelease>', self.fun_search)
-        var.bind('<Return>', self.fun_search)
-        var.bind('<KP_Enter>', self.fun_search)
-        var = ttk.Button(frm, text='Search', command=self.fun_search, width=10)
-        var.pack(side=tk.TOP)
-
-        line = ttk.Frame(frm)
-        line.pack(side=tk.TOP)
-        var = ttk.Checkbutton(line, variable=self.search_matchcase, text='Case')
-        var.pack(side=tk.LEFT)
-        var = ttk.Checkbutton(line, variable=self.search_wholeword, text='Word')
-        var.pack(side=tk.LEFT)
-
-    def ini_expression(self, frame: tk.Misc) -> tk.Text:
-        frm = ttk.Frame(frame)
-        frm.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-
-        var = ttk.Entry(frm, textvariable=self.expression_box)
-        var.pack(side=tk.TOP, fill=tk.X, expand=tk.YES)
-        # var.bind('<KeyRelease>', self.fun_expression_reset)
-        var.bind('<Return>', self.fun_expression)
-        var.bind('<KP_Enter>', self.fun_expression)
-
-        var = ttk.Label(frm, textvariable=self.expression_path)
-        var.pack(side=tk.TOP, expand=tk.YES, fill=tk.X)
-
-        var = ttk.Button(frm, text='Evaluate Expression', command=self.fun_expression)
-        var.pack(side=tk.TOP, fill=tk.X, expand=tk.YES)
-
-        text = tk.Text(frm, wrap=tk.NONE, width=DETAILS_TAB_WIDTH)
-        text.pack(fill=tk.BOTH, expand=tk.YES)
-
-        var = tk.Scrollbar(frm, orient=tk.HORIZONTAL, command=text.xview)
-        var.pack(side=tk.BOTTOM, fill=tk.X)
-        text.configure(xscrollcommand=var.set)
-        return text
-
-    "======================================================"
-    "================ general functions ==================="
-    "======================================================"
-
-    def check_expand(self):
-        open_close_all_tree(self.hdf_tree, "", self.expandall.get())
-
-    def _delete_tree(self):
-        self.hdf_tree.delete()
-        self.hdf_map.delete()
-
-    def populate_from_file(self, event=None):
-        filename = self.filepath.get()
-        self.map = hdfmap.create_nexus_map(filename)
-        with hdfmap.load_hdf(filename) as hdf:
-            self.populate(hdf, self.map)
-
-    def populate(self, hdf_obj: h5py.File, hdf_map: hdfmap.NexusMap):
-        self._delete_tree()
-        self.hdf_tree.populate(hdf_obj, openstate=self.expandall.get())
-        self.hdf_map.populate(hdf_obj, hdf_map)
-
-    "======================================================"
-    "================= event functions ===================="
-    "======================================================"
-
-    def tab_change(self, event):
-        # Selected tab:
-        selected = self.view_tabs.select()
-        tab_name = self.view_tabs.tab(selected)['text']
-        if tab_name == 'NeXus':
-            self.nexus.populate(self.map)
-        if tab_name == 'Tree String':
-            self.hdf_text.populate(self.map.filename)
-        if tab_name == 'NeXus2SRS':
-            self.nexus2srs.populate(self.map)
-        if tab_name == 'NXtransformations':
-            self.transformations.populate(self.map.filename)
-
-    def tree_select(self, event):
-        self.text.delete('1.0', tk.END)
-        tree = event.widget
-        item = tree.focus()
-        if 'path' in tree['columns']:  # hdfmap tab
-            path = tree.set(item, column='path')
-        else:
-            path = tree.item(item, 'text')
-        out = hdfobj_string(self.filepath.get(), path)
-        self.text.insert('1.0', out)
-
-    def select_file(self, event=None):
-        filename = select_hdf_file(self.root)
-        if filename:
-            self.filepath.set(filename)
-            self.populate_from_file()
-
-    def fun_search(self, event=None):
-        """Search currently active tab"""
-        if self.view_tabs.index(self.view_tabs.select()) == 0:
-            self.hdf_tree.tree.selection_remove(self.hdf_tree.tree.selection())
-            search_tree(
-                treeview=self.hdf_tree.tree,
-                branch="",
-                query=self.search_box.get(),
-                match_case=self.search_matchcase.get(),
-                whole_word=self.search_wholeword.get()
-            )
-        else:
-            self.hdf_map.tree.selection_remove(self.hdf_map.tree.selection())
-            search_tree(
-                treeview=self.hdf_map.tree,
-                branch="",
-                query=self.search_box.get(),
-                match_case=self.search_matchcase.get(),
-                whole_word=self.search_wholeword.get()
-            )
-
-    def fun_expression(self, event=None):
-        if self.map is None:
-            return
-        # self.text2.delete('1.0', tk.END)
-        expression = self.expression_box.get()
-        self.expression_path.set(f"path = {self.map.get_path(expression)}")
-        out_str = f">>> {expression}\n"
-        try:
-            # out = hdfmap.hdf_eval(self.filepath.get(), expression)
-            out = self.map.eval(self.map.load_hdf(), expression)
-        except NameError as ne:
-            out = ne
-        out_str += f"{out}\n\n"
-        # self.text2.insert('1.0', out_str)
-        self.text2.insert(tk.END, out_str)
