+++ mmg_toolbox/tkguis/widgets/scan_range_selector.py
@@ -24,11 +24,9 @@ class ScanRangeSelector:
 
         # variables
         self.exp_folder = tk.StringVar(self.root, initial_directory)
-        self.save_folder = tk.StringVar(self.root, '')
         self.number_start = tk.StringVar(self.root, '-10')
         self.number_end = tk.StringVar(self.root, '-1')
         self.number_step = tk.IntVar(self.root, 1)
-        self.metadata_name = tk.StringVar(self.root, '')
         self.file_list = []
 
         # Range selection
@@ -49,7 +47,7 @@ class ScanRangeSelector:
         var.bind("<Return>", self.update_numbers)
         ttk.Button(frm, text='Get numbers', command=self.numbers_from_exp).pack(side=tk.LEFT)
         ttk.Button(frm, text='Generate', command=self.update_numbers).pack(side=tk.LEFT, padx=4)
-        ttk.Button(frm, text='Select Files', command=self.select_files).pack(side=tk.RIGHT, padx=4)
+        ttk.Button(frm, text='Select Files', command=self.select_files).pack(side=tk.LEFT, padx=4)
 
         # Text box
         frm = ttk.Frame(self.root)
@@ -62,10 +60,7 @@ class ScanRangeSelector:
         var = ttk.Scrollbar(frm, orient=tk.VERTICAL, command=self.text.yview)
         var.pack(side=tk.LEFT, fill=tk.Y)
         self.text.configure(yscrollcommand=var.set)
-
-        frm = ttk.Frame(self.root)
-        frm.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-        ttk.Button(frm, text='Check', command=self.show_metadata).pack()
+        ttk.Button(frm, text='Check', command=self.show_metadata).pack(side=tk.LEFT, fill=tk.Y)
 
     def numbers_from_exp(self):
         exp_folder = self.exp_folder.get()
@@ -98,6 +93,7 @@ class ScanRangeSelector:
 
     def generate_scan_numbers(self) -> list[int]:
         scan_text = self.text.get("1.0", tk.END)
+        # TODO: replace eval with asteval
         scan_numbers = eval(scan_text)
         return scan_numbers
 
@@ -112,14 +108,16 @@ class ScanRangeSelector:
         return self.numbers2files(scan_numbers)
 
     def select_files(self):
-        from ..apps.scans import select_scans
-        files = select_scans(self.exp_folder.get(), self.root, self.config)
+        from ..apps.scans import list_scans
+        file_list = list_scans(self.exp_folder.get())
+        files = list_scans(*file_list, parent=self.root, config=self.config, button_name='Select')
         if files:
             numbers = [get_scan_number(file) for file in files]
             self.text.replace("1.0", tk.END, str(numbers))
 
     def show_metadata(self):
         from ..apps.scans import list_scans
-        metadata_list = self.metadata_name.get().split(',')
         file_list = self.generate_scan_files()
-        list_scans(*file_list, parent=self.root, config=self.config, metadata_list=metadata_list)
+        if len(file_list) == 0:
+            return
+        list_scans(*file_list, parent=self.root, config=self.config)
