+++ mmg_toolbox/tkguis/widgets/script_runner.py
@@ -6,12 +6,14 @@ import os
 import tkinter as tk
 from tkinter import ttk
 
-from ...scripts import scripts
-from mmg_toolbox.utils.env_functions import (get_scan_numbers, get_last_scan_number, get_first_file)
-from mmg_toolbox.utils.file_functions import get_scan_number, replace_scan_number
+import hdfmap
+
+from mmg_toolbox.scripts import scripts
+from mmg_toolbox.utils.env_functions import get_first_file, run_python_script
 from ..misc.logging import create_logger
 from ..misc.config import get_config
 from ..misc.functions import select_folder
+from ..widgets.scan_range_selector import ScanRangeSelector
 
 logger = create_logger(__file__)
 
@@ -70,37 +72,7 @@ class ScriptRunner:
         sec = ttk.LabelFrame(self.root, text='Scan Numbers')
         sec.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES, padx=4, pady=4)
 
-        frm = ttk.Frame(sec)
-        frm.pack(side=tk.TOP, fill=tk.BOTH, expand=tk.YES)
-
-        ttk.Label(frm, text='First:').pack(side=tk.LEFT, padx=2)
-        var = ttk.Entry(frm, textvariable=self.number_start, width=8)
-        var.pack(side=tk.LEFT, padx=2)
-        var.bind("<Return>", self.update_numbers)
-        ttk.Label(frm, text='Last:').pack(side=tk.LEFT, padx=2)
-        var = ttk.Entry(frm, textvariable=self.number_end, width=8)
-        var.pack(side=tk.LEFT, padx=2)
-        var.bind("<Return>", self.update_numbers)
-        ttk.Label(frm, text='Step:').pack(side=tk.LEFT, padx=2)
-        var = ttk.Entry(frm, textvariable=self.number_step, width=4)
-        var.pack(side=tk.LEFT, padx=2)
-        var.bind("<Return>", self.update_numbers)
-        ttk.Button(frm, text='Get numbers', command=self.numbers_from_exp).pack(side=tk.LEFT)
-        ttk.Button(frm, text='Generate', command=self.update_numbers).pack(side=tk.LEFT, padx=4)
-        ttk.Button(frm, text='Select Files', command=self.select_files).pack(side=tk.RIGHT, padx=4)
-
-        # Text box
-        frm = ttk.Frame(sec)
-        frm.pack(side=tk.TOP, expand=tk.YES, fill=tk.BOTH)
-
-        ttk.Label(frm, text='Scans = ').pack(side=tk.LEFT, padx=2)
-        self.text = tk.Text(frm, wrap=tk.WORD, width=65, height=5)
-        self.text.pack(side=tk.LEFT, fill=tk.BOTH, expand=tk.YES)
-
-        var = ttk.Scrollbar(frm, orient=tk.VERTICAL, command=self.text.yview)
-        var.pack(side=tk.LEFT, fill=tk.Y)
-        self.text.configure(yscrollcommand=var.set)
-        ttk.Button(frm, text='Check', command=self.show_metadata).pack(side=tk.LEFT, fill=tk.Y)
+        self.range = ScanRangeSelector(sec, exp_directory, self.config)
 
         # Script Selection
         sec = ttk.LabelFrame(self.root, text='Script')
@@ -124,69 +96,34 @@ class ScriptRunner:
 
         line = ttk.Frame(self.root)
         line.pack(side=tk.TOP, fill=tk.X, expand=tk.YES)
-        ttk.Label(line, text='file', width=10).pack(side=tk.LEFT, padx=2)
+        ttk.Label(line, text='file', width=6).pack(side=tk.LEFT, padx=2)
         ttk.Entry(line, textvariable=self.output_file, width=60).pack(side=tk.LEFT, padx=2)
         ttk.Button(line, text='RUN', command=self.run_template).pack(side=tk.LEFT)
+        ttk.Button(line, text='View', command=self.view_script).pack(side=tk.LEFT)
 
     def browse_metadata(self):
-        pass
+        from ..apps.namespace_select import create_metadata_selector
+        scan_file = next(iter(self.range.generate_scan_files()), get_first_file(self.exp_folder.get()))
+        hdf_map = hdfmap.create_nexus_map(scan_file)
+        paths = create_metadata_selector(hdf_map)
+        if paths:
+            self.metadata_name.set(', '.join(path for path in paths))
 
-    def numbers_from_exp(self):
-        exp_folder = self.exp_folder.get()
-        if exp_folder:
-            numbers = get_scan_numbers(exp_folder)
-            self.number_start.set(str(numbers[0]))
-            self.number_end.set(str(numbers[-1]))
-
-    def update_numbers(self, event=None):
-        first = eval(self.number_start.get())
-        last = eval(self.number_end.get())
-        step = self.number_step.get()
-
-        if (last - first) / step > 1000:
-            raise IOError('Range is too large')
-
-        exp_folder = self.exp_folder.get()
-        if exp_folder and (first < 1 or last < 1):
-            last_scan = get_last_scan_number(exp_folder)
-            if first < 1:
-                first = last_scan + first
-            if last < 1:
-                last = last_scan + last
-            # numbers = get_scan_numbers(exp_folder)
-            # if first < numbers[0] and last < numbers[0]:
-            #     scan_range = [numbers[idx] for idx in range(first, last+1, step)]
-
-        scan_range = list(range(first, last+1, step))
-        self.text.replace("1.0", tk.END, str(scan_range))
-
-    def generate_scan_numbers(self) -> list[int]:
-        scan_text = self.text.get("1.0", tk.END)
-        scan_numbers = eval(scan_text)
-        return scan_numbers
-
-    def numbers2files(self, scan_numbers: list[int]) -> list[str]:
-        exp_folder = self.exp_folder.get()
-        scan_file_template = get_first_file(exp_folder)
-        scan_files = (replace_scan_number(scan_file_template, number) for number in scan_numbers)
-        return [file for file in scan_files if os.path.isfile(file)]
-
-    def generate_scan_files(self) -> list[str]:
-        scan_numbers = self.generate_scan_numbers()
-        return self.numbers2files(scan_numbers)
-
-    def select_files(self):
-        from ..apps.scans import select_scans
-        files = select_scans(self.exp_folder.get(), self.root, self.config)
-        if files:
-            numbers = [get_scan_number(file) for file in files]
-            self.text.replace("1.0", tk.END, str(numbers))
-
-    def show_metadata(self):
-        from ..apps.scans import list_scans
-        metadata_list = self.metadata_name.get().split(',')
-        file_list = self.generate_scan_files()
-        list_scans(*file_list, parent=self.root, config=self.config, metadata_list=metadata_list)
+    def update_options(self):
+        new = {
+            # {{template}}: replacement
+            'beamline': self.config.get('beamline', ''),
+            # 'description': '',
+            'filepaths': '\n    '.join(self.range.generate_scan_files()),
+            'experiment_dir': self.exp_folder.get(),
+            'scan_numbers': str(self.range.generate_scan_numbers()),
+            # 'title': 'a nice plot',
+            'x-axis': 'axes',
+            'y-axis': 'signal',
+            'value': self.metadata_name.get()
+        }
+        self.options.clear()
+        self.options.update(new)
 
     def script_select(self, event=None):
         filename, desc = scripts.SCRIPTS[self.script_name.get()]
@@ -201,6 +138,9 @@ class ScriptRunner:
         self.output_file.set(os.path.join(proc, filename))
 
     def run_template(self, event=None):
+        if self.range.text.get("1.0", tk.END).strip() == '':
+            return
+        self.update_options()
         output_file = self.output_file.get()
         if output_file.endswith('.ipynb'):
             print(f"Creating notebook file: {output_file}")
@@ -212,14 +152,19 @@ class ScriptRunner:
             script_template = self.script_name.get()
             scripts.create_script(output_file, script_template, **self.options)
             print(f"Running script...")
+            run_python_script(output_file)
         else:
             raise Exception('File is not script or notebook')
 
-    def run_script(self, script_file: str, *scan_files: str, **options):
+    def run_script(self, script_file: str):
         pass
 
     def run_notebook(self, notebook_file: str, *scan_files: str, **options):
         pass
+
+    def view_script(self):
+        from ..apps.python_editor import create_python_editor
+        create_python_editor(open(self.output_file.get()).read(), parent=self.root, config=self.config)
 
     def browse_datadir(self):
         folder = select_folder(self.root)
@@ -230,3 +175,4 @@ class ScriptRunner:
         folder = select_folder(self.root)
         if folder:
             self.proc_folder.set(folder)
+
